requirements.txt
# libraries
pandas
numpy
matplotlib
plotly
scipy
seaborn
scikit-learn
tensorflow
keras
ta
dotenv

# api
yfinance
alpaca-py
# arch
arch
